Refactored endpoints to add and remove citizens (plus deleted an old unused file)
