--- GirafAPI/Utils/Role.cs
@@ -1,7 +1,0 @@
-namespace GirafAPI.Utils;
-
-public enum Role
-{
-    Administrator,
-    Trustee
-}
